df_openapi/urls.py
from django.conf import settings
from django.conf.urls.static import static
from django.contrib import admin
from django.urls import include
from django.urls import path
from drf_spectacular.views import SpectacularAPIView
from drf_spectacular.views import SpectacularRedocView
from drf_spectacular.views import SpectacularSwaggerView


urlpatterns = [
    path("admin/", admin.site.urls),

    path("api/v1/chat/", include("df_chat.drf.urls")),
    path("api/v1/auth/", include("df_auth.drf.urls")),
    path("api/v1/notifications/", include("df_notifications.drf.urls")),

    path("api/v1/schema/", SpectacularAPIView.as_view(), name="schema"),
    path("api/v1/", SpectacularRedocView.as_view(url_name="schema"), name="redoc"),
    path(
        "api/v1/swagger/",
        SpectacularSwaggerView.as_view(url_name="schema"),
        name="swagger-ui",
    ),
] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
